+++ web/src/components/organisms/Project/Content/ContentList/index.tsx
@@ -1,8 +1,6 @@
-import styled from "@emotion/styled";
 import { useCallback } from "react";
 import { useParams, useNavigate } from "react-router-dom";
 
-import Content from "@reearth-cms/components/atoms/Content";
 import ContentTableWrapper from "@reearth-cms/components/molecules/Content/ContentTableWrapper";
 import ModelsMenu from "@reearth-cms/components/organisms/Project/ModelsMenu";
 import { useT } from "@reearth-cms/i18n";
@@ -26,34 +24,13 @@ const ContentList: React.FC = () => {
   );
 
   return (
-    <PaddedContent>
-      <StyledModelsMenu title={t("Content")} selectModel={selectModel} />
-      <ContentChild>
-        <ContentTableWrapper
-          modelName={currentModel?.name}
-          contentTableFields={contentTableFields}
-          contentTableColumns={contentTableColumns}
-        />
-      </ContentChild>
-    </PaddedContent>
+    <ContentTableWrapper
+      modelName={currentModel?.name}
+      contentTableFields={contentTableFields}
+      contentTableColumns={contentTableColumns}>
+      <ModelsMenu title={t("Content")} selectModel={selectModel} />
+    </ContentTableWrapper>
   );
 };
-
-const PaddedContent = styled(Content)`
-  margin: 16px;
-  background-color: #fff;
-  display: flex;
-  min-height: 100%;
-`;
-
-const StyledModelsMenu = styled(ModelsMenu)`
-  width: 200px;
-`;
-
-const ContentChild = styled.div`
-  flex: 1;
-  background-color: #fff;
-  padding: 24px;
-`;
 
 export default ContentList;
